Improved error handling in tests
Before this change we weren't calling done() when there was an error,
which meant that we started running the next test before the first one
was finished thanks to Node's asynchronous-by-default nature.

With this we always call done(), sometimes with an error and sometimes
without, which means the tests run in series and fail more quickly when
the response comes back as an error.