Add support for OpenCage Geocoder
The OpenCage Geocoder is another geocoding service that can be used for
reverse geocoding. https://geocoder.opencagedata.com/

It requires a key which should be provided in the config options as
simply 'key'.

Signing up to OpenCage and getting a key is free. On the free tier users
can make 2500 requests per day.

In order to test the OpenCage integration set the OPENCAGE_KEY
environment variable.